% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\docType{data}
\name{bank}
\alias{bank}
\title{Bank Data set}
\format{
A data frame with with 100 rows and 2 columns:
\describe{
\item{arrival_time}{time customers arrives}
\item{service_time}{time it takes to service a customer}
...
}
}
\source{
on iLearn
}
\usage{
bank
}
\description{
'bank' will return a pair arrival and service times
}
\keyword{datasets}
