% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Multiserver}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{Number of customer arrivals.}

\item{ServiceTimes}{How long the service will take for the customer}

\item{NumServers}{How many servers there will be. For this case, it will be set to 1}
}
\value{
table data
}
\description{
This function will stimulate how customers will go through a first come, first serve queuing system as long as 'fate' has already decided when each customer will arrive and their service times.
}
